optimize uint8 quant tool 0x01
