+++ tools/quantize/quantization_uint8_perlayer/quant_save_graph.hpp
@@ -22,8 +22,8 @@
  * Author: hhchen@openailab.com
  */
 
-#include <stdlib.h>
-#include <stdio.h>
+#include <cstdlib>
+#include <cstdio>
 #include <sys/stat.h>
 #include <dirent.h>
